refactor: generalize getButton
